Merge main

+++ src/main/java/legend/game/types/BattleUiParts.java
@@ -222,11 +222,11 @@ public class BattleUiParts {
   }
 
   /** Also includes +, S, P as indices 10, 11, 12 */
-  public void queueBigNumber(final int digit, final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
+  public void queueBigNumber(final int digit, final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
     this.queue(this.bigNumberVert + digit * 4, x, y, 8, 16, packedClut, translucency, brightness, widthScale, heightScale);
   }
 
-  public void queuePoints(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
+  public void queuePoints(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
     this.queue(this.pointsVert, x, y, 8, 16, packedClut, translucency, brightness, widthScale, heightScale);
   }
 
@@ -242,11 +242,11 @@ public class BattleUiParts {
       overlay.heightScale_06 / (float)0x1000 + 1.0f);
   }
 
-  public void queueLetter(final int index, final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
+  public void queueLetter(final int index, final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
     this.queue(this.textVert + index * 4, x, y, 12, 12, packedClut, translucency, brightness, widthScale, heightScale);
   }
 
-  public void queueButton(final int index, final int x, final int y, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
+  public void queueButton(final int index, final float x, final float y, @Nullable final Translucency translucency, final int brightness, final float widthScale, final float heightScale) {
     final ButtonPressHudMetrics06 metrics = buttonPressHudMetrics_800faaa0[index];
 
     int w = metrics.wOrRightU_03;
@@ -260,48 +260,48 @@ public class BattleUiParts {
     this.queue(this.buttonVert + index * 4, x, y, w, h, metrics.packedClut_05, translucency, brightness, widthScale, heightScale);
   }
 
-  public void queueDaddyFrame(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyFrame(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyFrameVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
     this.queue(this.daddyFrameVert + 4, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyArrow(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyArrow(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyArrowVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyDarkEye(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyDarkEye(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyDarkEyeVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyFlatCenter(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyFlatCenter(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyFlatCenterVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyMeter(final int index, final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyMeter(final int index, final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyMeterVert + index * 4, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyEyeFlash(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyEyeFlash(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyEyeFlashVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyDivineIris(final int x, final int y) {
+  public void queueDaddyDivineIris(final float x, final float y) {
     this.queue(this.daddyDivineIrisVert, x, y, 0, 0, 0xc, Translucency.B_PLUS_F, 0x80, 0.5f, 1.5f);
   }
 
-  public void queueDaddyStar(final int x, final int y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
+  public void queueDaddyStar(final float x, final float y, final int packedClut, @Nullable final Translucency translucency, final int brightness) {
     this.queue(this.daddyStarVert, x, y, 0, 0, packedClut, translucency, brightness, 1.0f, 1.0f);
   }
 
-  public void queueDaddyButtonGlow(final int x, final int y, final int brightness, final float widthScale, final float heightScale) {
+  public void queueDaddyButtonGlow(final float x, final float y, final int brightness, final float widthScale, final float heightScale) {
     this.queue(this.daddyButtonGlowVert, x, y, 23, 23, 0xc, Translucency.B_PLUS_F, brightness, widthScale, heightScale);
   }
 
-  public void queueDaddyPerfect(final int index, final int x, final int y, final int brightness) {
+  public void queueDaddyPerfect(final int index, final float x, final float y, final int brightness) {
     this.queue(this.daddyPerfectVert + index * 4, x, y, 8, 16, 0x29, Translucency.B_PLUS_F, brightness, 1.0f, 1.0f);
   }
 
-  public void queue(final int vertexIndex, final int x, final int y, final int w, final int h, final int packedClut, @Nullable final Translucency transMode, final int brightness, final float widthScale, final float heightScale) {
+  public void queue(final int vertexIndex, final float x, final float y, final float w, final float h, final int packedClut, @Nullable final Translucency transMode, final int brightness, final float widthScale, final float heightScale) {
     final float left = x + w / 2.0f;
     final float top = y + h / 2.0f;
     final float offsetX = w * widthScale / 2.0f;
